feat: auth polish, UX improvements y cantidad editable en carrito
## Auth (Tasks 1 & 2)
- Alinea LoginResponseDto al contrato real del backend (campo 'tipo' en lugar de 'type', añade expiraEn, nombre, rol)
- Persiste username y displayName en DataStore a través de SessionManager
- AuthRepository llama a saveUserInfo() tras login satisfactorio
- Sidebar muestra nombre real del usuario autenticado y botón de logout (ExitToApp)
- LoginViewModel.parseError() distingue 400 (credenciales incorrectas) de errores 5xx
- AuthInterceptor excluye /api/auth/login de la limpieza de sesión en 401

## Mejoras incrementales (Task 3)
- Añade unidadVenta a ArticuloDto, CatalogProduct y ArticuloMapper
- unitLabelFor() / unitLabel(): UNIDAD→ud., CAJA→caja, GRANEL→granel, PESO→kg
- ProductCard (CatalogScreen) y ProductOrderCard (OrderBuilder) muestran "€X.XX / ud."
- Búsqueda extendida en OrderBuilderScreen y EditOrderScreen: nombre, descripción, codigoInterno y codigoBarras
- Botón de lector Bluetooth (QrCodeScanner) con foco automático y banner temporal de 3 s
- Elimina badge de pedidos de ClientCard (reducción de ruido visual)

## Cantidad editable en carrito (Task 4)
- CartItem.quantity: Int → Double (soporta decimales y negativos para devoluciones)
- QuantityControl reemplazado por EditableQuantityControl: BasicTextField entre [-] y [+]
  - Acepta enteros, decimales (coma o punto) y valores negativos
  - Estados intermedios (vacío, '-') preservados mientras el campo está enfocado
  - Al perder foco: valores inválidos restauran la cantidad previa; válidos hacen commit vía onSetQuantity
  - ImeAction.Done limpia el foco y dispara la validación
- OrderSummaryPanel.totalItems: Int → Double; display via formatQty()
- Totales negativos en línea de carrito mostrados en rojo
- EditLineRow (EditOrderScreen) aplica el mismo patrón con BasicTextField estilizado
- EditOrderViewModel: decrementLine ya no borra la línea al llegar a 0; nuevo setLineCantidad()
- CheckoutScreen: ReviewItemRow formatea Double con formatQty() (sin '.0' en enteros)

+++ app/src/main/java/com/example/distridulce/network/mapper/PedidoMapper.kt
@@ -36,7 +36,7 @@ fun List<CartItem>.toPedidoRequest(
             ?: return@mapNotNull null   // skip unmappable items
         PedidoLineaRequestDto(
             articuloId = articuloId,
-            cantidad   = item.quantity.toDouble()
+            cantidad   = item.quantity
         )
     }
